requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
pdfkit==1.0.0
matplotlib==3.8.2
httpx==0.25.1
python-multipart==0.0.6
pydantic==2.4.2 
wkhtmltopdf==0.2
PyPDF2==3.0.1
psycopg2-binary==2.9.9